+++ client/src/App.js
@@ -7,7 +7,6 @@ import { Alert } from "reactstrap";
 import NavBar from "./components/NavBar";
 import Add from "./components/Add";
 import Edit from "./components/Edit";
-import AdvancedSearch from './components/AdvancedSearch';
 import Search from "./components/Search";
 import NoCoordsSearch from "./components/NoCoordsSearch";
 import DirectoryAddUpdate from "./components/DirectoryAddUpdate";
@@ -39,7 +38,6 @@ function App() {
               <Switch>
                 <Route exact path="/" component={Add} />
                 <Route path="/add" component={Add} />
-                <Route path="/advanced" component={AdvancedSearch} />
                 <Route path="/edit/:id" component={Edit} />
                 <Route path="/search" component={Search} />
                 <Route path="/nocoords" component={NoCoordsSearch} />
